feat: MySQL와 연동 완료
src 폴더 하위 db.js 파일 생성하여 mysql과 서버 연동
