--- app.js
@@ -1,12 +1,0 @@
-import express from 'express';
-const app = express();
-
-app.set('port', process.env.PORT || 3000);
-
-app.get('/', (req, res) => {
-  res.send('Hello, Express');
-});
-
-app.listen(app.get('port'), () => {
-  console.log(app.get('port'), '번 포트에서 대기 중');
-});
