+++ package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "app.js",
   "scripts": {
-    "start": "babel-node app.js",
-    "dev": "nodemon --exec babel-node app.js",
+    "start": "babel-node src/app.js",
+    "dev": "nodemon --exec babel-node src/app.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -14,7 +14,9 @@
   "dependencies": {
     "@babel/node": "^7.22.6",
     "@babel/preset-env": "^7.22.9",
-    "express": "^4.18.2"
+    "dotenv": "^16.3.1",
+    "express": "^4.18.2",
+    "mysql": "^2.18.1"
   },
   "devDependencies": {
     "@babel/core": "^7.22.9",
